move example to examples
