+++ python/examples/adapter_google_adk/pyproject.toml
@@ -1,12 +1,11 @@
 [project]
 name = "thirdweb-ai-examples-autogen"
 version = "0.1.0"
-description = "Example of using thirdweb_ai with AutoGen"
+description = "Example of using thirdweb_ai with Google ADK"
 authors = [{ name = "thirdweb", email = "support@thirdweb.com" }]
 requires-python = "~=3.10"
 dependencies = [
-    "thirdweb-ai[google-adk]",
-    "google-adk>=1.0.0",
+    "thirdweb-ai[google-adk]"
 ]
 
 [tool.uv]
